refactor: 타임테이블 정보 type 파일 분리
