+++ apps/client/src/pages/time-table/components/stage/stage.tsx
@@ -1,34 +1,6 @@
 import * as styles from './stage.css';
-interface Artist {
-  artistId: string;
-  artistName: string;
-}
-
-interface FestivalTimes {
-  festivalTimeId: number;
-  startAt: string;
-  endAt: string;
-  isSelected: boolean;
-  artists: Artist[];
-}
-
-interface Stage {
-  stageOrder: number;
-  stageName: string;
-  festivalTimes: FestivalTimes[];
-}
-
-interface TimeTableInfo {
-  ticketOpenAt: string;
-  stageCount: number;
-  stages: Stage[];
-}
-
-interface StageProps {
-  timeTableInfo: TimeTableInfo;
-}
-
-const Stage = ({ timeTableInfo }: StageProps) => {
+import { TimeTableInfoType } from '@pages/time-table/types/time-table-info-type';
+const Stage = ({ timeTableInfo }: TimeTableInfoType) => {
   return (
     <div className={styles.stageWrapper}>
       {timeTableInfo.stages.map(({ stageOrder, stageName }) => (
